Move generic validation to trait CanValidate
